Crate collect

collect-rs is intended as an experimental extension of the Rust standard library's libcollections. Ideas that are too niche, crazy, or experimental to land in libcollections can be gathered here where they can gain the maintenance and network-effect benefits that libcollections enjoys, but without worrying about such ivory tower concepts as "general usefulness" and "consistency".

For the time being we plan to be highly volatile with a low barrier of entry. We want to explore the space of data structuring in Rust. We want to prove out ideas and implementations that could one day make their way into the standard library.

Got a Concurrent, Immutable, or Persistent collection? Awesome! Crazy ideas for collection or iterator adapters? Heck yeah!

Come on in!

Note that anything include in collect-rs is theoretically a candidate for inclusion in libcollections. As such, this project is licensed under the same terms as Rust itself.

Reexports

pub use blist::BList;

pub use enum_set::EnumSet;

pub use immut_slist::ImmutSList;

pub use interval_heap::IntervalHeap;

pub use lru_cache::LruCache;

Modules

blist

compare
Comparators.

enum_set
A structure for holding a set of enum variants.

immut_slist

interval_heap
A double-ended priority queue implemented with an interval heap.

iter

lru_cache
A cache that holds a limited number of key-value pairs. When the capacity of the cache is exceeded, the least-recently-used (where "used" means a look-up or putting the pair into the cache) pair is automatically removed.

proto

tree_map

tree_set

trie_map

trie_set

Structs

TreeMap
This is implemented as an AA tree, which is a simplified variation of a red-black tree where red (horizontal) nodes can only be added as a right child. The time complexity is the same, and re-balancing operations are more frequent but also cheaper.

TreeSet
An implementation of a set on top of the TreeMap container.

TrieMap
A map implemented as a radix trie.

TrieSet
A set implemented as a radix trie.