Struct collect::blist::BList

pub struct BList<T> {
    // some fields omitted
}

A skeleton implementation of a BList, based on the Space-Efficient Linked List described in Open Data Structures.

A BList is a hybrid between an array and a doubly-linked-list. It consists of arrays in a doubly-linked-list. In this way we get many of the nice properties of a DList, but with improved cache properties and less allocations.

A BList's B-factor is analogous to the same factor in a BTree. It guarantees that all nodes contain B-1 and B+1 elements (except the ends). Once a position has been identified to perform an insertion or deletion, it will take amortized O(B) time to perform, with a worst-case cost of O(B^2). Insertion and deletion on either end will always take O(1) time, though (assuming it takes O(1) time to allocate an array of size B).

Methods

impl<T> BList<T>

fn new() -> BList<T>

Creates a new BList with a reasonable choice for B.

fn with_b(b: usize) -> BList<T>

Creates a new BList with the specified B.

impl<T> BList<T>

fn push_back(&mut self, elem: T)

Inserts the element at the back of the list.

fn push_front(&mut self, elem: T)

Inserts the element at the front of the list.

fn pop_back(&mut self) -> Option<T>

Removes and returns an element off the back of the list. Returns None if empty.

fn pop_front(&mut self) -> Option<T>

Removes and returns an element off the front of the list. Returns None if empty.

fn front(&self) -> Option<&T>

Gets an immutable reference to the first element in the list.

fn back(&self) -> Option<&T>

Gets an immutable reference to the last element in the list.

fn front_mut(&mut self) -> Option<&mut T>

Gets a mutable reference to the first element in the list.

fn back_mut(&mut self) -> Option<&mut T>

Gets a mutable reference to the last element in the list.

fn len(&self) -> usize

Gets the number of elements in the list.

fn is_empty(&self) -> bool

Returns true if the list contains no elements, or false otherwise.

fn clear(&mut self)

Drops everything in the list.

fn iter(&self) -> Iter<T>

Gets a by-reference iterator over the elements in the list.

fn iter_mut(&mut self) -> IterMut<T>

Gets a by-mutable-reference iterator over the elements in the list.

fn into_iter(self) -> IntoIter<T>

Gets a by-value iterator over the elements in the list.

fn append_lazy(&mut self, other: &mut BList<T>)

Lazily moves the contents of other to the end of self, in the sense that it makes no effort to preserve the node-size lower-bound invariant. This can have negative effects on the effeciency of the resulting list, but is otherwise much faster than a proper invariant-preserving append.

Panics

Panics if the lists have a different value of B.