Struct collect::compare::Extract

pub struct Extract<E, C, T: ?Sized, K> where E: Fn(&T) -> K, C: Compare<K> {
    // some fields omitted
}

A comparator that extracts a sort key from a value.

Examples

use collect::compare::{Compare, Extract, Natural};
use std::cmp::Ordering::Greater;

let a = vec![1, 2, 3];
let b = vec![4, 5];

let cmp = Extract::new(|vec: &Vec<u8>| vec.len(), Natural);
assert_eq!(cmp.compare(&a, &b), Greater);

Methods

impl<E, C, T: ?Sized, K> Extract<E, C, T, K> where E: Fn(&T) -> K, C: Compare<K>

fn new(ext: E, cmp: C) -> Extract<E, C, T, K>

Returns a comparator that extracts a sort key using ext and compares it using cmp.