Struct collect::immut_slist::ImmutSList

pub struct ImmutSList<T> {
    // some fields omitted
}

An immutable singly-linked list, as seen in basically every functional language

Methods

impl<T> ImmutSList<T>

fn new() -> ImmutSList<T>

Constructs a new, empty ImmutSList

fn append(&self, elem: T) -> ImmutSList<T>

Returns a copy of the list, with elem appended to the front

fn head(&self) -> Option<&T>

Returns a reference to the first element in the list

fn tail(&self) -> ImmutSList<T>

Returns a copy of the list, with the first element removed

fn tailn(&self, n: usize) -> ImmutSList<T>

Returns a copy of the list, with the first n elements removed

fn last(&self) -> Option<&T>

Returns the last element in the list

fn lastn(&self, n: usize) -> ImmutSList<T>

Returns a copy of the list, with only the last n elements remaining

fn iter<'a>(&'a self) -> Iter<'a, T>

Returns an iterator over references to the elements of the list in order

fn len(&self) -> usize

fn is_empty(&self) -> bool