Module collect::interval_heap

A double-ended priority queue implemented with an interval heap.

Structs

IntervalHeap
An IntervalHeap is an implementation of a double-ended priority queue. As such, it supports the following operations: push, get_min, get_max, pop_min, pop_max where insertion takes amortized O(log n) time, removal takes O(log n) time and accessing minimum and maximum can be done in constant time. Also, other convenient functions are provided that handle conversion from and into vectors and allow iteration etc.

Iter
IntervalHeap iterator.