Struct collect::interval_heap::IntervalHeap

pub struct IntervalHeap<T, C: Compare<T> = Natural<T>> {
    // some fields omitted
}

An IntervalHeap is an implementation of a double-ended priority queue. As such, it supports the following operations: push, get_min, get_max, pop_min, pop_max where insertion takes amortized O(log n) time, removal takes O(log n) time and accessing minimum and maximum can be done in constant time. Also, other convenient functions are provided that handle conversion from and into vectors and allow iteration etc.

Methods

impl<T: Ord> IntervalHeap<T>

fn new() -> IntervalHeap<T>

Returns an empty heap ordered according to the natural order of its elements.

Examples

use collect::IntervalHeap;

let heap = IntervalHeap::<u32>::new();
assert!(heap.is_empty());

fn with_capacity(capacity: usize) -> IntervalHeap<T>

Returns an empty heap with the given capacity and ordered according to the natural order of its elements.

The heap will be able to hold exactly capacity elements without reallocating.

Examples

use collect::IntervalHeap;

let heap = IntervalHeap::<u32>::with_capacity(5);
assert!(heap.is_empty());
assert!(heap.capacity() >= 5);

fn from_vec(vec: Vec<T>) -> IntervalHeap<T>

Returns a heap containing all the elements of the given vector and ordered according to the natural order of its elements.

Examples

use collect::IntervalHeap;

let heap = IntervalHeap::from_vec(vec![5, 1, 6, 4]);
assert_eq!(heap.len(), 4);
assert_eq!(heap.get_min_max(), Some((&1, &6)));

impl<T, C: Compare<T>> IntervalHeap<T, C>

fn with_comparator(cmp: C) -> IntervalHeap<T, C>

Returns an empty heap ordered according to the given comparator.

fn with_capacity_and_comparator(capacity: usize, cmp: C) -> IntervalHeap<T, C>

Returns an empty heap with the given capacity and ordered according to the given comparator.

fn from_vec_and_comparator(vec: Vec<T>, cmp: C) -> IntervalHeap<T, C>

Returns a heap containing all the elements of the given vector and ordered according to the given comparator.

fn iter(&self) -> Iter<T>

An iterator visiting all values in underlying vector, in arbitrary order.

fn get_min(&self) -> Option<&T>

Returns a reference to the smallest item or None (if empty).

fn get_max(&self) -> Option<&T>

Returns a reference to the greatest item or None (if empty).

fn get_min_max(&self) -> Option<(&T, &T)>

Returns references to the smallest and greatest item or None (if empty).

fn capacity(&self) -> usize

Returns the number of items the interval heap could hold without reallocation.

fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity for exactly additional more elements to be inserted in the given IntervalHeap. Does nothing if the capacity is already sufficient.

Note that the allocator may give the collection more space than it requests. Therefore capacity can not be relied upon to be precisely minimal. Prefer reserve if future insertions are expected.

fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the IntervalHeap. The collection may reserve more space to avoid frequent reallocations.

fn shrink_to_fit(&mut self)

Discards as much additional capacity as possible.

fn pop_min(&mut self) -> Option<T>

Removes the smallest item and returns it, or None if is empty.

fn pop_max(&mut self) -> Option<T>

Removes the greatest item and returns it, or None if is empty.

fn push(&mut self, x: T)

Pushes an item onto the queue.

fn into_vec(self) -> Vec<T>

Consumes the IntervalHeap and returns the underlying vector in arbitrary order.

fn into_sorted_vec(self) -> Vec<T>

Consumes the IntervalHeap and returns a vector in sorted (ascending) order.

fn len(&self) -> usize

Returns the number of items in the interval heap

fn is_empty(&self) -> bool

Returns true if the queue contains no items.

fn clear(&mut self)

Drops all items from the queue.