Trait collect::iter::OrderedMapIterator

pub trait OrderedMapIterator<K, A>: Iterator<Item=(K, A)> + Sized {
    fn inner_join_map<B, T: OrderedMapIterator<K, B>>(self, map: T) -> InnerJoinMapIterator<Self, T> { ... }
    fn inner_join_set<B, T: OrderedSetIterator<K>>(self, set: T) -> InnerJoinMapSetIterator<Self, T> { ... }
    fn outer_join<B, T: OrderedMapIterator<K, B>>(self, right: T) -> OuterJoinIterator<Self, T, (K, A), (K, B)> { ... }
}

Allows an iterator to be do an inner join with another iterator to combine their values or filter based on their keys. this trait is applied to an iterator over a map like structure

Provided Methods

fn inner_join_map<B, T: OrderedMapIterator<K, B>>(self, map: T) -> InnerJoinMapIterator<Self, T>

join two ordered maps together

fn inner_join_set<B, T: OrderedSetIterator<K>>(self, set: T) -> InnerJoinMapSetIterator<Self, T>

filter an ordered map with an ordered set

fn outer_join<B, T: OrderedMapIterator<K, B>>(self, right: T) -> OuterJoinIterator<Self, T, (K, A), (K, B)>

Join an ordered iterator with the right ordered iterator. The new iterator will return a key value pair for every key in either iterator. If a key is present in both iterators they will be returned together (two values). If a value is in the Right, but not the left iterator it will be return without the value in the left iterator. If the value is in the left iterator by not the right that will be return without the value from the left iterator.