Trait collect::iter::OrderedSetIterator

pub trait OrderedSetIterator<K>: Iterator + Sized {
    fn inner_join_map<B, T: OrderedMapIterator<K, B>>(self, map: T) -> InnerJoinMapSetIterator<T, Self> { ... }
    fn inner_join_set<T: OrderedSetIterator<K>>(self, map: T) -> InnerJoinSetIterator<Self, T> { ... }
}

Allows an iterator to be do an inner join with another iterator to combine their values or filter based on their keys. this trait is applied to an iterator over a set like structure

Provided Methods

fn inner_join_map<B, T: OrderedMapIterator<K, B>>(self, map: T) -> InnerJoinMapSetIterator<T, Self>

join two ordered maps together

fn inner_join_set<T: OrderedSetIterator<K>>(self, map: T) -> InnerJoinSetIterator<Self, T>

filter an ordered map with an ordered set