Module collect::lru_cache

A cache that holds a limited number of key-value pairs. When the capacity of the cache is exceeded, the least-recently-used (where "used" means a look-up or putting the pair into the cache) pair is automatically removed.

Example

use collect::LruCache;

let mut cache: LruCache<int, int> = LruCache::new(2);
cache.insert(1, 10);
cache.insert(2, 20);
cache.insert(3, 30);
assert!(cache.get(&1).is_none());
assert_eq!(*cache.get(&2).unwrap(), 20);
assert_eq!(*cache.get(&3).unwrap(), 30);

cache.insert(2, 22);
assert_eq!(*cache.get(&2).unwrap(), 22);

cache.insert(6, 60);
assert!(cache.get(&3).is_none());

cache.set_capacity(1);
assert!(cache.get(&2).is_none());

Structs

LruCache
An LRU Cache.