Struct collect::lru_cache::LruCache

pub struct LruCache<K, V> {
    // some fields omitted
}

An LRU Cache.

Methods

impl<K: Hash<HmHasher> + Eq, V> LruCache<K, V>

fn new(capacity: usize) -> LruCache<K, V>

Create an LRU Cache that holds at most capacity items.

Example

use collect::LruCache;
let mut cache: LruCache<i32, &str> = LruCache::new(10);

fn put(&mut self, k: K, v: V)

Deprecated: Replaced with insert.

fn insert(&mut self, k: K, v: V) -> Option<V>

Inserts a key-value pair into the cache. If the key already existed, the old value is returned.

Example

use collect::LruCache;
let mut cache = LruCache::new(2);

cache.insert(1, "a");
cache.insert(2, "b");
assert_eq!(cache.get(&1), Some(&"a"));
assert_eq!(cache.get(&2), Some(&"b"));

fn get(&mut self, k: &K) -> Option<&V>

Return a value corresponding to the key in the cache.

Example

use collect::LruCache;
let mut cache = LruCache::new(2);

cache.insert(1, "a");
cache.insert(2, "b");
cache.insert(2, "c");
cache.insert(3, "d");

assert_eq!(cache.get(&1), None);
assert_eq!(cache.get(&2), Some(&"c"));

fn pop(&mut self, k: &K) -> Option<V>

Deprecated: Renamed to remove.

fn remove(&mut self, k: &K) -> Option<V>

Remove and return a value corresponding to the key from the cache.

Example

use collect::LruCache;
let mut cache = LruCache::new(2);

cache.insert(2, "a");

assert_eq!(cache.remove(&1), None);
assert_eq!(cache.remove(&2), Some("a"));
assert_eq!(cache.remove(&2), None);
assert_eq!(cache.len(), 0);

fn capacity(&self) -> usize

Return the maximum number of key-value pairs the cache can hold.

Example

use collect::LruCache;
let mut cache: LruCache<i32, &str> = LruCache::new(2);
assert_eq!(cache.capacity(), 2);

fn change_capacity(&mut self, capacity: usize)

Deprecated: Renamed to set_capacity.

fn set_capacity(&mut self, capacity: usize)

Change the number of key-value pairs the cache can hold. Remove least-recently-used key-value pairs if necessary.

Example

use collect::LruCache;
let mut cache = LruCache::new(2);

cache.insert(1, "a");
cache.insert(2, "b");
cache.insert(3, "c");

assert_eq!(cache.get(&1), None);
assert_eq!(cache.get(&2), Some(&"b"));
assert_eq!(cache.get(&3), Some(&"c"));

cache.set_capacity(3);
cache.insert(1, "a");
cache.insert(2, "b");

assert_eq!(cache.get(&1), Some(&"a"));
assert_eq!(cache.get(&2), Some(&"b"));
assert_eq!(cache.get(&3), Some(&"c"));

cache.set_capacity(1);

assert_eq!(cache.get(&1), None);
assert_eq!(cache.get(&2), None);
assert_eq!(cache.get(&3), Some(&"c"));

fn len(&self) -> usize

Return the number of key-value pairs in the cache.

fn is_empty(&self) -> bool

Returns whether the cache is currently empty.

fn clear(&mut self)

Clear the cache of all key-value pairs.