Struct collect::proto::par_vec::ParVec

pub struct ParVec<T> {
    // some fields omitted
}

A vector that can be operated on concurrently via non-overlapping slices.

Get a ParVec and a vector of slices via new(), send the slices to other threads and mutate them, then get the mutated vector with into_inner() when finished.

Methods

impl<T: Send + Sync> ParVec<T>

fn new(vec: Vec<T>, slices: uint) -> (ParVec<T>, Vec<ParSlice<T>>)

Create a new ParVec, returning it and a vector of slices that can be sent to other threads and mutated concurrently.

fn into_inner_opt(self) -> Result<Vec<T>, ParVec<T>>

Take the inner Vec if there are no slices remaining. Returns Err(self) if there are still slices out there.

fn into_inner(self) -> Vec<T>

Take the inner Vec, waiting until all slices have been freed.