Module collect::tree_set

Structs

Difference
A lazy iterator producing elements in the set difference (in-order).

Intersection
A lazy iterator producing elements in the set intersection (in-order).

IntoIter
A lazy forward iterator over a set that consumes the set while iterating.

Iter
A lazy forward iterator over a set.

RevIter
A lazy backward iterator over a set.

SymmetricDifference
A lazy iterator producing elements in the set symmetric difference (in-order).

TreeSet
An implementation of a set on top of the TreeMap container.

Union
A lazy iterator producing elements in the set union (in-order).