Trait collect::compare::Compare

pub trait Compare<Lhs: ?Sized, Rhs: ?Sized = Lhs> {
    fn compare(&self, lhs: &Lhs, rhs: &Rhs) -> Ordering;

    fn compares_lt(&self, lhs: &Lhs, rhs: &Rhs) -> bool { ... }
    fn compares_le(&self, lhs: &Lhs, rhs: &Rhs) -> bool { ... }
    fn compares_ge(&self, lhs: &Lhs, rhs: &Rhs) -> bool { ... }
    fn compares_gt(&self, lhs: &Lhs, rhs: &Rhs) -> bool { ... }
    fn compares_eq(&self, lhs: &Lhs, rhs: &Rhs) -> bool { ... }
    fn compares_ne(&self, lhs: &Lhs, rhs: &Rhs) -> bool { ... }
}

A comparator imposing a total order.

See the compare module's documentation for detailed usage.

The compares_* methods may be overridden to provide more efficient implementations.

Required Methods

fn compare(&self, lhs: &Lhs, rhs: &Rhs) -> Ordering

Compares two values, returning Less, Equal, or Greater if lhs is less than, equal to, or greater than rhs, respectively.

Provided Methods

fn compares_lt(&self, lhs: &Lhs, rhs: &Rhs) -> bool

Checks if lhs is less than rhs.

fn compares_le(&self, lhs: &Lhs, rhs: &Rhs) -> bool

Checks if lhs is less than or equal to rhs.

fn compares_ge(&self, lhs: &Lhs, rhs: &Rhs) -> bool

Checks if lhs is greater than or equal to rhs.

fn compares_gt(&self, lhs: &Lhs, rhs: &Rhs) -> bool

Checks if lhs is greater than rhs.

fn compares_eq(&self, lhs: &Lhs, rhs: &Rhs) -> bool

Checks if lhs is equal to rhs.

fn compares_ne(&self, lhs: &Lhs, rhs: &Rhs) -> bool

Checks if lhs is not equal to rhs.